requirements.txt
absl-py==0.13.0
anyio==3.2.1
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
bleach==3.3.1
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==4.2.2
captum==0.4.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.0
click==8.0.1
cloudpickle==1.6.0
coverage==5.5
cryptography==3.4.7
cycler==0.10.0
cytoolz==0.11.0
dask==2021.7.0
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
dtcwt==0.12.0
entrypoints==0.3
fsspec==2021.7.0
google-auth==1.33.0
google-auth-oauthlib==0.4.4
GPUtil==1.4.0
grpcio==1.38.1
idna==3.1
imagecodecs==2021.6.8
imageio==2.9.0
importlib-metadata==4.6.1
iniconfig==1.1.1
ipykernel==6.0.2
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.9.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.1
kiwisolver==1.3.1
locket==0.2.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.1
nltk==3.6.2
notebook==6.4.0
numexpr==2.7.3
numpy==1.21.0
oauthlib==3.1.1
olefile==0.46
packaging==21.0
pandas==1.3.0
pandocfilters==1.4.2
parso==0.8.2
partd==1.2.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pip==21.1.3
pluggy==0.13.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
py==1.10.0
py3nvml==0.2.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyCoShREM==0.0.2
pycparser==2.20
Pygments==2.9.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
pytest==6.2.4
pytest-cov==2.12.1
python-dateutil==2.8.2
pytorch-wavelets==1.3.0
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
scikit-image==0.18.1
scipy==1.7.0
Send2Trash==1.7.1
setuptools==49.6.0.post20210108
sip==4.19.13
six==1.16.0
sniffio==1.2.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.10.1
testpath==0.5.0
tifffile==2021.4.8
toml==0.10.2
toolz==0.11.1
torch==1.9.0
torchattacks==3.0.0
torchsummary==1.5.1
torchvision==0.10.0
tornado==6.1
tqdm==4.62.2
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==2.0.1
wheel==0.36.2
xmltodict==0.12.0
zipp==3.5.0
